// IamFuture-Dex-Frontend/src/config/constants/types.ts
export type PageMeta = {
  title: string
  description?: string
  image?: string
}